superhero-image-classification/app/requirements.txt
Pillow
numpy
ultralytics
streamlit